Move court tests to own file
