+++ judgments/tests/tests.py
@@ -1,15 +1,14 @@
 import re
 from os import environ
 from unittest import skip
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 from dateutil import parser as dateparser
 from django.test import TestCase
 from lxml import etree
 
 from judgments import converters, utils
-from judgments.models import CourtDates, SearchResult, SearchResults
-from judgments.templatetags.court_utils import get_court_date_range, get_court_name
+from judgments.models import SearchResult, SearchResults
 from judgments.utils import as_integer, display_back_link, paginator
 
 
@@ -463,45 +462,3 @@ def test_solo_stop_word_regex():
     stop_words = ["and", "of", "the", "for"]
     expected_output = r"(^and$)|(^of$)|(^the$)|(^for$)"
     assert utils.solo_stop_word_regex(stop_words) == expected_output
-
-
-def test_get_court_name():
-    assert get_court_name("uksc") == "United Kingdom Supreme Court"
-
-
-def test_get_court_name_non_existent():
-    assert get_court_name("ffff") == ""
-
-
-@patch("judgments.templatetags.court_utils.CourtDates.objects.get")
-class TestCourtDatesHelper(TestCase):
-    def mock_court_dates(self, start_year, end_year):
-        mock = Mock()
-        mock.configure_mock(start_year=start_year, end_year=end_year)
-        return mock
-
-    def test_when_court_with_param_exists_and_no_dates_in_db_and_start_end_same(
-        self, get
-    ):
-        get.side_effect = CourtDates.DoesNotExist
-        court = self.mock_court_dates(2011, 2011)
-        self.assertEqual(get_court_date_range(court), "2011")
-
-    def test_when_court_with_param_exists_and_no_dates_in_db_and_start_end_different(
-        self, get
-    ):
-        get.side_effect = CourtDates.DoesNotExist
-        court = self.mock_court_dates(2011, 2012)
-        self.assertEqual(get_court_date_range(court), "2011 &ndash; 2012")
-
-    def test_when_court_with_param_exists_and_dates_in_db_and_start_end_same(self, get):
-        get.return_value = self.mock_court_dates(2013, 2013)
-        court = self.mock_court_dates(2011, 2012)
-        self.assertEqual(get_court_date_range(court), "2013")
-
-    def test_when_court_with_param_exists_and_dates_in_db_and_start_end_different(
-        self, get
-    ):
-        get.return_value = self.mock_court_dates(2013, 2015)
-        court = self.mock_court_dates(2011, 2012)
-        self.assertEqual(get_court_date_range(court), "2013 &ndash; 2015")
